% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat.R
\name{new_strat}
\alias{new_strat}
\title{Constructor for object of class "strat"}
\usage{
new_strat(z, X, base_strata = NULL, refined_strata = NULL, details = NULL)
}
\arguments{
\item{z}{Vector of treatment assignment}

\item{X}{Covariate matrix or data.frame}

\item{base_strata}{Original strata, if they exist}

\item{refined_strata}{Refined strata, if they exist}

\item{details}{A list of details from the linear program.
Also includes \code{X_std} if calculated}
}
\description{
Creates an object of S3 class "strat"
}
