% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat.R
\name{validate_strat}
\alias{validate_strat}
\title{Validator for object of class "strat"}
\usage{
validate_strat(object)
}
\arguments{
\item{object}{An object of class \code{strat}}
}
\value{
Error or object of class \code{strat} if valid
}
\description{
Checks validity of an object of S3 class "strat"
}
